+++ frontend/rust-lib/event-integration/Cargo.toml
@@ -19,7 +19,7 @@ flowy-encrypt = { workspace = true }
 lib-dispatch = { workspace = true }
 lib-infra = { workspace = true }
 flowy-server = { path = "../flowy-server" }
-flowy-server-config = { workspace = true }
+flowy-server-pub = { workspace = true }
 flowy-notification  = { workspace = true }
 anyhow.workspace = true
 flowy-storage = { workspace = true }
