+++ frontend/rust-lib/flowy-server/Cargo.toml
@@ -34,7 +34,7 @@ flowy-folder-pub = { workspace = true }
 flowy-database-pub = { workspace = true }
 flowy-document-pub = { workspace = true }
 flowy-error = { workspace = true, features = ["impl_from_postgres", "impl_from_serde", "impl_from_reqwest", "impl_from_url", "impl_from_appflowy_cloud"] }
-flowy-server-config = { workspace = true }
+flowy-server-pub = { workspace = true }
 flowy-encrypt = { workspace = true }
 flowy-storage = { workspace = true }
 mime_guess = "2.0"
