+++ frontend/rust-lib/flowy-server/src/supabase/file_storage/core.rs
@@ -10,7 +10,7 @@ use url::Url;
 
 use flowy_encrypt::{decrypt_data, encrypt_data};
 use flowy_error::FlowyError;
-use flowy_server_config::supabase_config::SupabaseConfiguration;
+use flowy_server_pub::supabase_config::SupabaseConfiguration;
 use flowy_storage::{FileStoragePlan, FileStorageService, StorageObject};
 use lib_infra::future::FutureResult;
