+++ frontend/rust-lib/flowy-user/src/manager.rs
@@ -12,7 +12,7 @@ use collab_integrate::collab_builder::AppFlowyCollabBuilder;
 use collab_integrate::CollabKVDB;
 use flowy_error::{internal_error, ErrorCode, FlowyResult};
 use flowy_folder_pub::entities::ImportData;
-use flowy_server_config::AuthenticatorType;
+use flowy_server_pub::AuthenticatorType;
 use flowy_sqlite::kv::StorePreferences;
 use flowy_sqlite::schema::user_table;
 use flowy_sqlite::ConnectionPool;
