+++ frontend/rust-lib/lib-infra/Cargo.toml
@@ -19,6 +19,8 @@ anyhow.workspace = true
 walkdir = "2.4.0"
 tempfile = "3.8.1"
 validator = "0.16.0"
+tracing.workspace = true
+atomic_refcell = "0.1"
 
 [dev-dependencies]
 rand = "0.8.5"
